LibGfx/JPEG: Move the `HuffmanTable` class to its own file
